completed contract

+++ programs/amm/src/constants.rs
@@ -2,4 +2,4 @@ use anchor_lang::prelude::*;
 
 #[constant]
 pub const SEED:&str = "amm";
-pub const ANCHOR_DISCRIMINATOR: u8 = 8;
+pub const ANCHOR_DISCRIMINATOR: usize = 8;
